Unfortunately, {bino,fib}heap can't be generic
Hence it's not being written. The nodes in the heap need
to store the key value for the heap to work, but that
would require the nodes themselves to be generic. This
makes it impossible to do nicely with C functions + macros.
Doing it all in macros also doesn't work as insert and
removemin need to be able to create new variables, which can
be done by just using them in the macros (causes name clashes)
or by allocating them ahead of time in the binoheap struct and
using them when needed (wasteful, one of them is a buffer).