Redid splat as a "templated" library.
I realized I was incorrect about binoheap, it is
possible to do a generic intrusive data structure
that needs a compare operation in C, it's just
really goddamn ugly. It's basically just templates
implemented with the preprocessor. The "templated"
functions exist simply because they could not be
written safely as macros, as there could be
potential name clashes.